Moved debounce to Scheduler to avoid need for implicitly passing schedulers
