+++ core/shared/src/main/scala/fs2/Scheduler.scala
@@ -114,6 +114,61 @@ abstract class Scheduler {
     sleep(d).drain
 
   /**
+   * Debounce the stream with a minimum period of `d` between each element.
+   *
+   * @example {{{
+   * scala> import scala.concurrent.duration._, scala.concurrent.ExecutionContext.Implicits.global, cats.effect.IO
+   * scala> val s2 = Scheduler[IO](1).flatMap { scheduler =>
+   *      |   val s = Stream(1, 2, 3) ++ scheduler.sleep_[IO](400.millis) ++ Stream(4, 5) ++ scheduler.sleep_[IO](10.millis) ++ Stream(6)
+   *      |   s.through(scheduler.debounce(200.milliseconds))
+   *      | }
+   * scala> s2.runLog.unsafeRunSync
+   * res0: Vector[Int] = Vector(3, 6)
+   * }}}
+   */
+  def debounce[F[_],O](d: FiniteDuration)(implicit F: Effect[F], ec: ExecutionContext): Pipe[F,O,O] = {
+    def unconsLatest(s: Stream[F,O]): Pull[F,Nothing,Option[(O,Stream[F,O])]] =
+      s.pull.uncons.flatMap {
+        case Some((hd,tl)) => Pull.segment(hd.last).flatMap {
+          case (_, Some(last)) => Pull.pure(Some(last -> tl))
+          case (_, None) => unconsLatest(tl)
+        }
+        case None => Pull.pure(None)
+      }
+
+    def go(o: O, s: Stream[F,O]): Pull[F,O,Unit] = {
+      sleep[F](d).pull.unconsAsync.flatMap { l =>
+        s.pull.unconsAsync.flatMap { r =>
+          (l race r).pull.flatMap {
+            case Left(_) =>
+              Pull.output1(o) >> r.pull.flatMap {
+                case Some((hd,tl)) => Pull.segment(hd.last).flatMap {
+                  case (_, Some(last)) => go(last, tl)
+                  case (_, None) => unconsLatest(tl).flatMap {
+                    case Some((last, tl)) => go(last, tl)
+                    case None => Pull.done
+                  }
+                }
+                case None => Pull.done
+              }
+            case Right(r) => r match {
+              case Some((hd,tl)) => Pull.segment(hd.last).flatMap {
+                case (_, Some(last)) => go(last, tl)
+                case (_, None) => go(o, tl)
+              }
+              case None => Pull.output1(o)
+            }
+          }
+        }
+      }
+    }
+    in => unconsLatest(in).flatMap {
+      case Some((last,tl)) => go(last, tl)
+      case None => Pull.done
+    }.stream
+  }
+
+  /**
    * Returns an execution context that executes all tasks after a specified delay.
    */
   def delayedExecutionContext(delay: FiniteDuration, reporter: Throwable => Unit = ExecutionContext.defaultReporter): ExecutionContext = new ExecutionContext {
